move list_glob func to the iterator
